+++ Base/Xrpl.BinaryCodec/Enums/Field.AccountId.Generated.cs
@@ -28,6 +28,8 @@ namespace Xrpl.BinaryCodec.Enums
         public static readonly AccountIdField Borrower = new AccountIdField(nameof(Borrower), 25);
         public static readonly AccountIdField Counterparty = new AccountIdField(nameof(Counterparty), 26);
         public static readonly AccountIdField Sponsor = new AccountIdField(nameof(Sponsor), 27);
+        public static readonly AccountIdField HighSponsor = new AccountIdField(nameof(HighSponsor), 28);
+        public static readonly AccountIdField LowSponsor = new AccountIdField(nameof(LowSponsor), 29);
         public static readonly AccountIdField CounterpartySponsor = new AccountIdField(nameof(CounterpartySponsor), 30);
         public static readonly AccountIdField Sponsee = new AccountIdField(nameof(Sponsee), 31);
     }