+++ Xrpl/Models/Ledger/LOPayChannel.cs
@@ -103,5 +103,13 @@ namespace Xrpl.Models.Ledger
         //todo not found field DestinationNode?: string
         //A hint indicating which page of the destination's owner directory links to
         //this object, in case the directory consists of multiple pages.
-    }
+    
+        /// <summary>Sequence (or ticket) of the PaymentChannelCreate that created this channel.</summary>
+        [JsonPropertyName("Sequence")]
+        public uint? Sequence { get; set; }
+
+        /// <summary>Destination directory page hint (hex UInt64).</summary>
+        [JsonPropertyName("DestinationNode")]
+        public string DestinationNode { get; set; }
+}
 }